Experimental version of a recording mechanism for OMPT events on the Intel Xeon Phi using the Intel COI library
This commit adds a highly experimental recording mechanism for OMPT events on the Intel Xeon Phi. The OMPT events are recorded per thread on the Xeon Phi in a buffer. A buffer is transferred to the host as soon as it is full or the device runtime shuts down.

Issues:
- It is not possible to record ompt_event_thread_end and ompt_event_runtime_shutdown, because these events are signaled with the COI process destruction. There seems to be no communication between host and device possible when those events occur.
- The maximum number of buffers is statically set to 240.
- The transfer of data between host and device uses global variables on device side.
- Stopping tracing is currently not implemented.