Move ompt_event_task_begin after the task has been fully created
After it is allocated in __kmp_omp_task_alloc, the code generated by the
compiler will fill in the firstprivates and addresses of the shareds.
So it is only fully created when it is brought to execution in the
different functions (depending on untied and whether there are dependencies).

Conflicts:
	runtime/src/kmp_tasking.c