Redesign of the OMPT Tracing API.
This redesign has to goals:

(1) Improve the stability. With the former version we had deadlocks from time to time. By using one request and one truncation event per thread this can be avoided (and the tests I did lock very promising right now). Furthermore, we use one separated thread for for the request and truncation signaling now. However, since we do not know how many threads will be started on the host, we have strict thread limitation of 240 at the moment. This has to be fixed later.

(2) Improve the performance. With one event per thread, we can avoid the global locks on the device. This patch is a preparation for removing these locks (but still not done with this patch)